Wait 3 seconds before exiting on ES connection error.
IOError gets thrown when the uploader can't do the initial connect to
the ES cluster. This may mean that we are just bringing up the services
after a system bounce or whatever, and ES hasn't had the chance to come
up yet. Waiting a bit before exiting helps supervisor - which doesn't
have a backoff time - to wait a bit before trying to bring the service
up again.